Move documentation contracts to Repl.Documentation namespace
