+++ src/Repl.Core/Documentation/ReplDocApp.cs
@@ -1,4 +1,4 @@
-namespace Repl;
+namespace Repl.Documentation;
 
 /// <summary>
 /// Application metadata.
