+++ src/Repl.Core/Documentation/ReplDocArgument.cs
@@ -1,4 +1,4 @@
-namespace Repl;
+namespace Repl.Documentation;
 
 /// <summary>
 /// Argument metadata.
