+++ src/Repl.Core/Documentation/ReplDocCommand.cs
@@ -1,4 +1,4 @@
-namespace Repl;
+namespace Repl.Documentation;
 
 /// <summary>
 /// Command metadata.
