+++ src/Repl.Core/Documentation/ReplDocContext.cs
@@ -1,4 +1,4 @@
-namespace Repl;
+namespace Repl.Documentation;
 
 /// <summary>
 /// Context metadata.
