+++ src/Repl.Core/Documentation/ReplDocumentationModel.cs
@@ -1,4 +1,4 @@
-namespace Repl;
+namespace Repl.Documentation;
 
 /// <summary>
 /// Structured documentation payload for the command graph.
