+++ src/Repl.Core/GlobalUsings.cs
@@ -1,1 +1,2 @@
 global using Repl.Parameters;
+global using Repl.Documentation;
