+++ src/Repl.Core/ReplDocumentationExtensions.cs
@@ -1,3 +1,5 @@
+using Repl.Documentation;
+
 namespace Repl;
 
 /// <summary>
