+++ src/Repl.Defaults/ReplAppDocumentationExtensions.cs
@@ -1,3 +1,5 @@
+using Repl.Documentation;
+
 namespace Repl;
 
 /// <summary>
